refactor: move form types from entities to shared
