+++ src/entities/Form/index.ts
@@ -1,13 +1,4 @@
 'use client';
 
 export { Form } from './ui';
-export type {
-  TBookingFormParams,
-  TSigninParams,
-  TSignupParams,
-  TRecoverPasswordParams,
-  TCallbackParams,
-  TNewPassword,
-} from './typing';
-
 export { type TFieldConfig } from './Field';
