+++ src/entities/Form/typing.ts
@@ -1,6 +1,8 @@
 import { ReactNode } from 'react';
 import { FieldValues } from 'react-hook-form';
 
+import { type TFormData } from '@/shared/form';
+
 import { TFieldConfig } from './Field';
 
 export type TForm = {
@@ -10,43 +12,3 @@ export type TForm = {
   defaultValues?: FieldValues;
   asterisk?: boolean;
 };
-
-export type TBookingFormParams = {
-  daterange: number[];
-  message: string;
-};
-
-export type TSigninParams = {
-  login: string;
-  password: string;
-};
-
-export type TSignupParams = {
-  lastName: string;
-  firstName: string;
-  middleName: string;
-  phone: string;
-  email: string;
-  password: string;
-  passwordRepeat: string;
-};
-
-export type TRecoverPasswordParams = {
-  email: string;
-};
-
-export type TNewPassword = {
-  token: string;
-  password: string;
-};
-
-export type TCallbackParams = {
-  phone: string;
-  comment: string;
-};
-
-export type TFormData = TBookingFormParams &
-  TSigninParams &
-  TSignupParams &
-  TRecoverPasswordParams &
-  TCallbackParams;
