+++ src/shared/form/index.ts
@@ -1,2 +1,12 @@
-export type { TReactHookFormInputElement, TKindOfField } from './typing';
+export type {
+  TReactHookFormInputElement,
+  TKindOfField,
+  TFormData,
+  TBookingFormParams,
+  TSigninParams,
+  TSignupParams,
+  TRecoverPasswordParams,
+  TNewPassword,
+  TCallbackParams,
+} from './typing';
 export { validationTypes, validationPattern } from './constants';
