+++ src/shared/form/typing.ts
@@ -1,10 +1,10 @@
 import { UseFormRegisterReturn, UseFormSetValue, UseFormWatch } from 'react-hook-form';
 
 export type TReactHookFormInputElement = {
-  name?: string;
+  name?: keyof TFormData;
   register?: UseFormRegisterReturn;
-  watch?: UseFormWatch<Record<string, string>>;
-  setValue?: UseFormSetValue<Record<string, string>>;
+  watch?: UseFormWatch<TFormData>;
+  setValue?: UseFormSetValue<TFormData>;
 };
 
 export type TKindOfField = 'input' | 'textarea' | 'checkbox' | 'daterange' | 'passwordInput';
@@ -13,3 +13,44 @@ export type TValidation = {
   value: RegExp;
   message: string;
 };
+
+export type TBookingFormParams = {
+  daterange: number[];
+  message: string;
+};
+
+export type TSigninParams = {
+  login: string;
+  password: string;
+};
+
+export type TSignupParams = {
+  lastName: string;
+  firstName: string;
+  middleName: string;
+  phone: string;
+  email: string;
+  password: string;
+  passwordRepeat: string;
+};
+
+export type TRecoverPasswordParams = {
+  email: string;
+};
+
+export type TNewPassword = {
+  token: string;
+  password: string;
+};
+
+export type TCallbackParams = {
+  phone: string;
+  comment: string;
+};
+
+export type TFormData = TBookingFormParams &
+  TSigninParams &
+  TSignupParams &
+  TRecoverPasswordParams &
+  TNewPassword &
+  TCallbackParams;
